Readd Roact, add Rodux!
Much better data handling after this. I was initially reluctant to use Rodux but it's pretty powerful for this circumstance.
