+++ package.json
@@ -10,19 +10,20 @@
   "author": "",
   "license": "MIT",
   "devDependencies": {
+    "@rbxts/cubic-bezier": "^1.0.2",
+    "@rbxts/roact": "^1.3.0-ts.6",
+    "@rbxts/roact-rodux": "^0.1.2",
+    "@rbxts/rodux": "^1.1.0",
     "@rbxts/services": "^1.1.2",
-    "@rbxts/types": "^1.0.420",
+    "@rbxts/types": "^1.0.428",
     "@typescript-eslint/eslint-plugin": "^4.8.0",
     "@typescript-eslint/parser": "^4.8.0",
     "eslint": "^7.13.0",
     "eslint-config-prettier": "^6.15.0",
     "eslint-plugin-prettier": "^3.1.4",
     "eslint-plugin-roblox-ts": "0.0.24",
-    "prettier": "^2.1.2"
-  },
-  "dependencies": {
-    "@rbxts/cubic-bezier": "^1.0.2",
-    "@rbxts/roact": "^1.3.0-ts.2",
+    "prettier": "^2.1.2",
     "typescript": "^4.0.3"
-  }
+  },
+  "dependencies": {}
 }
