+++ src/data/crash_config.json
@@ -7,6 +7,6 @@
     "Ouch.", "Never gonna give you up. Until now.", "Hello!", "Did I do well?", "All your Studio belong to us",
     "Don't panic! Alright? Stop panicking! I can still stop this. Oh. There's a password.", "Uhm. A...A...A? Nope?",
     "Witty comment available :)", "Ooh, shiny!", "Sir - mint cookies or graham crackers?", "I'm hurting, Dr. Studio",
-    "Does not use Java!", "War ich das?"],
+    "Fortunately does not use Java!", "Unfortunately does use JavaScript!", "War ich das?"],
   "format": "CRASH - %c"
 }
